Extract popover logic to dedicated mixin.
